perf: move async modules to individual files
