+++ src/routes/_actions/report.js
@@ -1,4 +1,4 @@
-import { importShowReportDialog } from '../_components/dialog/asyncDialogs'
+import { importShowReportDialog } from '../_components/dialog/asyncDialogs/importShowReportDialog.js'
 
 export async function reportStatusOrAccount ({ status, account }) {
   const showReportDialog = await importShowReportDialog()
