+++ src/routes/_components/Nav.html
@@ -50,7 +50,7 @@
 </style>
 <script>
   import NavItem from './NavItem'
-  import { importNavShortcuts } from '../_utils/asyncModules'
+  import { importNavShortcuts } from '../_utils/asyncModules/importNavShortcuts.js'
   import { store } from '../_store/store'
 
   export default {
