+++ src/routes/_components/NavShortcuts.html
@@ -19,7 +19,8 @@
 <script>
   import Shortcut from './shortcut/Shortcut'
   import { goto } from '../../../__sapper__/client'
-  import { importShowShortcutHelpDialog, importShowComposeDialog } from './dialog/asyncDialogs'
+  import { importShowShortcutHelpDialog } from './dialog/asyncDialogs/importShowShortcutHelpDialog'
+  import { importShowComposeDialog } from './dialog/asyncDialogs/importShowComposeDialog'
   import { store } from '../_store/store'
 
   export default {
