+++ src/routes/_components/compose/ComposeMediaItem.html
@@ -130,7 +130,7 @@
   import { ONE_TRANSPARENT_PIXEL } from '../../_static/media'
   import { get } from '../../_utils/lodash-lite'
   import { coordsToPercent } from '../../_utils/coordsToPercent'
-  import { importShowMediaEditDialog } from '../dialog/asyncDialogs'
+  import { importShowMediaEditDialog } from '../dialog/asyncDialogs/importShowMediaEditDialog.js'
   import { throttleTimer } from '../../_utils/throttleTimer'
 
   const updateMediaInStore = throttleTimer(scheduleIdleTask)
