+++ src/routes/_components/compose/ComposeStickyButton.html
@@ -31,7 +31,7 @@
 <script>
   import ComposeButton from './ComposeButton.html'
   import { store } from '../../_store/store'
-  import { importShowComposeDialog } from '../dialog/asyncDialogs'
+  import { importShowComposeDialog } from '../dialog/asyncDialogs/importShowComposeDialog.js'
   import { observe } from 'svelte-extras'
   import { classname } from '../../_utils/classname'
