+++ src/routes/_components/compose/ComposeToolbar.html
@@ -70,7 +70,8 @@
 <script>
   import IconButton from '../IconButton.html'
   import { store } from '../../_store/store'
-  import { importShowEmojiDialog, importShowPostPrivacyDialog } from '../dialog/asyncDialogs'
+  import { importShowEmojiDialog } from '../dialog/asyncDialogs/importShowEmojiDialog'
+  import { importShowPostPrivacyDialog } from '../dialog/asyncDialogs/importShowPostPrivacyDialog'
   import { doMediaUpload } from '../../_actions/media'
   import { toggleContentWarningShown } from '../../_actions/contentWarnings'
   import { mediaAccept } from '../../_static/media'
