+++ src/routes/_components/dialog/asyncDialogs/importShowAccountProfileOptionsDialog.js
@@ -1,0 +1,3 @@
+export const importShowAccountProfileOptionsDialog = () => import(
+  /* webpackChunkName: 'showAccountProfileOptionsDialog' */ '../creators/showAccountProfileOptionsDialog'
+).then(mod => mod.default)
