+++ src/routes/_components/dialog/asyncDialogs/importShowComposeDialog.js
@@ -1,0 +1,3 @@
+export const importShowComposeDialog = () => import(
+  /* webpackChunkName: 'showComposeDialog' */ '../creators/showComposeDialog'
+).then(mod => mod.default)
