+++ src/routes/_components/dialog/asyncDialogs/importShowCopyDialog.js
@@ -1,0 +1,3 @@
+export const importShowCopyDialog = () => import(
+  /* webpackChunkName: 'showCopyDialog' */ '../creators/showCopyDialog'
+).then(mod => mod.default)
