+++ src/routes/_components/dialog/asyncDialogs/importShowEmojiDialog.js
@@ -1,0 +1,3 @@
+export const importShowEmojiDialog = () => import(
+  /* webpackChunkName: 'showEmojiDialog' */ '../creators/showEmojiDialog'
+).then(mod => mod.default)
