+++ src/routes/_components/dialog/asyncDialogs/importShowMediaDialog.js
@@ -1,0 +1,3 @@
+export const importShowMediaDialog = () => import(
+  /* webpackChunkName: 'showMediaDialog' */ '../creators/showMediaDialog'
+).then(mod => mod.default)
