+++ src/routes/_components/dialog/asyncDialogs/importShowMediaEditDialog.js
@@ -1,0 +1,3 @@
+export const importShowMediaEditDialog = () => import(
+  /* webpackChunkName: 'showMediaEditDialog' */ '../creators/showMediaEditDialog'
+).then(mod => mod.default)
