+++ src/routes/_components/dialog/asyncDialogs/importShowMuteDialog.js
@@ -1,0 +1,3 @@
+export const importShowMuteDialog = () => import(
+  /* webpackChunkName: 'showMuteDialog' */ '../creators/showMuteDialog'
+).then(mod => mod.default)
