+++ src/routes/_components/dialog/asyncDialogs/importShowPostPrivacyDialog.js
@@ -1,0 +1,3 @@
+export const importShowPostPrivacyDialog = () => import(
+  /* webpackChunkName: 'showPostPrivacyDialog' */ '../creators/showPostPrivacyDialog'
+).then(mod => mod.default)
