+++ src/routes/_components/dialog/asyncDialogs/importShowReportDialog.js
@@ -1,0 +1,3 @@
+export const importShowReportDialog = () => import(
+  /* webpackChunkName: 'showReportDialog' */ '../creators/showReportDialog'
+).then(mod => mod.default)
