+++ src/routes/_components/dialog/asyncDialogs/importShowShortcutHelpDialog.js
@@ -1,0 +1,3 @@
+export const importShowShortcutHelpDialog = () => import(
+  /* webpackChunkName: 'showShortcutHelpDialog' */ '../creators/showShortcutHelpDialog'
+).then(mod => mod.default)
