+++ src/routes/_components/dialog/asyncDialogs/importShowTextConfirmationDialog.js
@@ -1,0 +1,3 @@
+export const importShowTextConfirmationDialog = () => import(
+  /* webpackChunkName: 'showTextConfirmationDialog' */ '../creators/showTextConfirmationDialog'
+).then(mod => mod.default)
