+++ src/routes/_components/profile/AccountProfileDetails.html
@@ -123,7 +123,7 @@
 </style>
 <script>
   import IconButton from '../IconButton.html'
-  import { importShowAccountProfileOptionsDialog } from '../dialog/asyncDialogs'
+  import { importShowAccountProfileOptionsDialog } from '../dialog/asyncDialogs/importShowAccountProfileOptionsDialog.js'
 
   const numberFormat = new Intl.NumberFormat('en-US')
