+++ src/routes/_components/profile/AccountProfileHeader.html
@@ -124,7 +124,7 @@
   import { removeEmoji } from '../../_utils/removeEmoji'
   import { store } from '../../_store/store'
   import Label from '../Label.html'
-  import { importShowMediaDialog } from '../dialog/asyncDialogs'
+  import { importShowMediaDialog } from '../dialog/asyncDialogs/importShowMediaDialog.js'
   import { getImageNativeDimensions } from '../../_utils/getImageNativeDimensions'
 
   export default {
