+++ src/routes/_components/settings/instance/InstanceActions.html
@@ -21,7 +21,7 @@
 </style>
 <script>
   import { store } from '../../../_store/store'
-  import { importShowTextConfirmationDialog } from '../../dialog/asyncDialogs'
+  import { importShowTextConfirmationDialog } from '../../dialog/asyncDialogs/importShowTextConfirmationDialog.js'
   import { switchToInstance, logOutOfInstance } from '../../../_actions/instances'
 
   export default {
