+++ src/routes/_components/settings/instance/PushNotificationSettings.html
@@ -41,7 +41,7 @@
 </style>
 <script>
   import { store } from '../../../_store/store'
-  import { importShowTextConfirmationDialog } from '../../dialog/asyncDialogs'
+  import { importShowTextConfirmationDialog } from '../../dialog/asyncDialogs/importShowTextConfirmationDialog.js'
   import { logOutOfInstance } from '../../../_actions/instances'
   import { updatePushSubscriptionForInstance, updateAlerts } from '../../../_actions/pushSubscription'
   import { toast } from '../../toast/toast'
