+++ src/routes/_components/status/Media.html
@@ -102,7 +102,7 @@
 </style>
 <script>
   import { DEFAULT_MEDIA_WIDTH, DEFAULT_MEDIA_HEIGHT, ONE_TRANSPARENT_PIXEL } from '../../_static/media'
-  import { importShowMediaDialog } from '../dialog/asyncDialogs'
+  import { importShowMediaDialog } from '../dialog/asyncDialogs/importShowMediaDialog.js'
   import { mouseover } from '../../_utils/events'
   import NonAutoplayGifv from '../NonAutoplayGifv.html'
   import PlayVideoIcon from '../PlayVideoIcon.html'
