+++ src/routes/_store/observers/setupLoggedInObservers.js
@@ -1,4 +1,4 @@
-import { importLoggedInStoreExtensions } from '../../_utils/asyncModules'
+import { importLoggedInStoreExtensions } from '../../_utils/asyncModules/importLoggedInStoreExtensions.js'
 
 // An observer that calls an observer... this is a bit weird, but it eliminates
 // circular dependencies and also allows us to lazy load observers/computations
