+++ src/routes/_utils/asyncModules/importArrowKeyNavigation.js
@@ -1,0 +1,3 @@
+export const importArrowKeyNavigation = () => import(
+  /* webpackChunkName: 'arrow-key-navigation' */ 'arrow-key-navigation'
+)
