+++ src/routes/_utils/asyncModules/importComposeBox.js
@@ -1,0 +1,3 @@
+export const importComposeBox = () => import(
+  /* webpackChunkName: 'ComposeBox.html' */ '../../_components/compose/ComposeBox.html'
+).then(mod => mod.default)
