+++ src/routes/_utils/asyncModules/importDatabase.js
@@ -1,0 +1,3 @@
+export const importDatabase = () => import(
+  /* webpackChunkName: 'database.js' */ '../../_database/databaseApis.js'
+)
