+++ src/routes/_utils/asyncModules/importEmojiMart.js
@@ -1,0 +1,3 @@
+export const importEmojiMart = () => import(
+  /* webpackChunkName: 'createEmojiMartPickerFromData.js' */ '../../_react/createEmojiMartPickerFromData.js'
+).then(mod => mod.default)
