+++ src/routes/_utils/asyncModules/importList.js
@@ -1,0 +1,3 @@
+export const importList = () => import(
+  /* webpackChunkName: 'List.html' */ '../../_components/list/List.html'
+).then(mod => mod.default)
