+++ src/routes/_utils/asyncModules/importLoggedInStoreExtensions.js
@@ -1,0 +1,3 @@
+export const importLoggedInStoreExtensions = () => import(
+  /* webpackChunkName: 'loggedInStoreExtensions.js' */ '../../_store/loggedInStoreExtensions.js'
+)
