+++ src/routes/_utils/asyncModules/importNavShortcuts.js
@@ -1,0 +1,3 @@
+export const importNavShortcuts = () => import(
+  /* webpackChunkName: 'NavShortcuts' */ '../../_components/NavShortcuts.html'
+).then(mod => mod.default)
