+++ src/routes/_utils/asyncModules/importNotificationVirtualListItem.js
@@ -1,0 +1,3 @@
+export const importNotificationVirtualListItem = () => import(
+  /* webpackChunkName: 'NotificationVirtualListItem.html' */ '../../_components/timeline/NotificationVirtualListItem.html'
+).then(mod => mod.default)
