+++ src/routes/_utils/asyncModules/importPageLifecycle.js
@@ -1,0 +1,3 @@
+export const importPageLifecycle = () => import(
+  /* webpackChunkName: 'page-lifecycle' */ 'page-lifecycle/dist/lifecycle.mjs'
+).then(mod => mod.default)
