+++ src/routes/_utils/asyncModules/importSnackbar.js
@@ -1,0 +1,3 @@
+export const importSnackbar = () => import(
+  /* webpackChunkName: 'Snackbar.html' */ '../../_components/snackbar/Snackbar.html'
+).then(mod => mod.default)
