+++ src/routes/_utils/asyncModules/importStatusVirtualListItem.js
@@ -1,0 +1,3 @@
+export const importStatusVirtualListItem = () => import(
+  /* webpackChunkName: 'StatusVirtualListItem.html' */ '../../_components/timeline/StatusVirtualListItem.html'
+).then(mod => mod.default)
