+++ src/routes/_utils/asyncModules/importTesseractWorker.js
@@ -1,0 +1,3 @@
+export const importTesseractWorker = () => import(
+  /* webpackChunkName: 'tesseractWorker' */ '../../_utils/tesseractWorker.js'
+).then(mod => mod.default)
