+++ src/routes/_utils/asyncModules/importTimeline.js
@@ -1,0 +1,3 @@
+export const importTimeline = () => import(
+  /* webpackChunkName: 'Timeline' */ '../../_components/timeline/Timeline.html'
+).then(mod => mod.default)
