+++ src/routes/_utils/asyncModules/importToast.js
@@ -1,0 +1,3 @@
+export const importToast = () => import(
+  /* webpackChunkName: 'Toast.html' */ '../../_components/toast/Toast.html'
+).then(mod => mod.default)
