+++ src/routes/_utils/asyncModules/importVirtualList.js
@@ -1,0 +1,3 @@
+export const importVirtualList = () => import(
+  /* webpackChunkName: 'VirtualList.html' */ '../../_components/virtualList/VirtualList.html'
+).then(mod => mod.default)
