+++ src/routes/_utils/asyncModules/importVirtualListStore.js
@@ -1,0 +1,3 @@
+export const importVirtualListStore = () => import(
+  /* webpackChunkName: 'virtualListStore.js' */ '../../_components/virtualList/virtualListStore.js'
+)
