+++ src/routes/share.html
@@ -3,7 +3,7 @@
   import { store } from './_store/store'
   import { goto } from '../../__sapper__/client'
   import { decodeURIComponentWithPluses } from './_utils/decodeURIComponentWithPluses'
-  import { importLoggedInStoreExtensions } from './_utils/asyncModules'
+  import { importLoggedInStoreExtensions } from './_utils/asyncModules/importLoggedInStoreExtensions.js'
 
   const SHARE_KEYS = ['title', 'text', 'url']
